Deduplicate spl/token and spl/token-2022 via shared common crate
Extract 7 identical source files (accounts, balances, memo, metadata,
mints, permissions, transfers) into a new spl/token-common crate.
Both spl/token and spl/token-2022 are now thin wrappers that only
define their program ID filter and delegate to token-common.

Closes #118
