+++ spl/token-2022/Cargo.toml
@@ -10,7 +10,5 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 substreams = { workspace = true }
 substreams-solana = { workspace = true }
-substreams-solana-program-instructions = { workspace = true }
-spl-token-metadata-interface = "0.7.0"
 proto = { path = "../../proto" }
-common = { path = "../../common" }
+spl-token-common = { path = "../token-common" }
