+++ spl/token/Cargo.toml
@@ -11,7 +11,5 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 substreams = { workspace = true }
 substreams-solana = { workspace = true }
-substreams-solana-program-instructions = { workspace = true }
-spl-token-metadata-interface = "0.7.0"
 proto = { path = "../../proto" }
-common = { path = "../../common" }
+spl-token-common = { path = "../token-common" }
